netcode refactor
