+++ core/src/contrib/systems/core/DamageSystem.ts
@@ -1,6 +1,5 @@
 import { Entity, Health, Position, SystemBuilder } from "@piggo-gg/core";
 
-
 export const DamageSystem: SystemBuilder<"DamageSystem"> = {
   id: "DamageSystem",
   init: ({ world }) => {
