+++ core/src/runtime/IsometricGame.ts
@@ -1,9 +1,9 @@
 import { AbilityHUD, Chat, ConnectButton, DebugButton, FullscreenButton, GameBuilder, Joystick, Noob, isMobile } from "@piggo-gg/core";
 
-export const IsometricGame = <T extends string>(g: GameBuilder<T>): GameBuilder<T> => ({
-  ...g,
+export const IsometricGame = <T extends string>(gameBuilder: GameBuilder<T>): GameBuilder<T> => ({
+  ...gameBuilder,
   init: (world) => {
-    const game = g.init(world);
+    const game = gameBuilder.init(world);
 
     if (world.runtimeMode === "client") {
       game.entities.push(Noob({ id: world.client!.playerId }));
