+++ core/src/runtime/IsometricWorld.ts
@@ -1,7 +1,7 @@
 import {
   ActionSystem, ClickableSystem, CommandSystem,
   CooldownSystem, Cursor, DamageSystem, DebugSystem,
-  GameCommand, InputSystem, NPCSystem, PhysicsSystem,
+  ExpiresSystem, GameCommand, InputSystem, NPCSystem, PhysicsSystem,
   RenderSystem, SpawnCommand, World, WorldBuilder, isMobile
 } from "@piggo-gg/core";
 
@@ -15,9 +15,9 @@ export const IsometricWorld: WorldBuilder = (props) => {
   if (!isMobile()) world.addEntity(Cursor());
 
   world.addSystemBuilders([
-    InputSystem, ClickableSystem, DebugSystem, DamageSystem,
-    CommandSystem, NPCSystem, CooldownSystem, ActionSystem,
-    PhysicsSystem, RenderSystem
+    ExpiresSystem, InputSystem, ClickableSystem, DebugSystem,
+    DamageSystem, CommandSystem, NPCSystem, CooldownSystem,
+    ActionSystem, PhysicsSystem, RenderSystem
   ]);
 
   return world;
